Login fixes + cucumber features

+++ app/controllers/user_sessions_controller.rb
@@ -1,6 +1,6 @@
 class UserSessionsController < ApplicationController
+  
   before_filter :require_no_user, :only => [:new, :create]
-  before_filter :require_user, :only => :destroy
   
   def new
     @user_session = UserSession.new
@@ -17,9 +17,11 @@ class UserSessionsController < ApplicationController
   end
   
   def destroy
-    current_user_session.destroy
-    flash[:notice] = "Logout successful!"
-    redirect_back_or_default new_user_session_url
+    if current_user
+      current_user_session.destroy
+      flash[:notice] = "Logout successful!"
+    end
+    redirect_to home_url
   end
 end
